Adding Sequences and Finger Trees
Adding Sequences as a specialization of the finger tree data structre.
These sequences support constant amortized access to both ends, aswell
as push'ing to both sides of the list. They also offer fast access to
an arbitrary index of the tree, and because of that are a nice
structure to use as random-access lists.